Refactor: GoogleServiceFactory 파일 분리
