+++ src/services/answer-refiner/answer-refiner.ts
@@ -8,8 +8,8 @@ import type {
 } from "../../types";
 
 import { GeminiClient } from "@/services/gemini-client";
+import { GoogleServiceFactory } from "@/services/google-service-factory";
 import { GoogleSheetsService } from "@/services/google-sheets";
-import { GoogleSheetsServiceFactory } from "@/services/index";
 
 export class AnswerRefiner {
   private geminiClient: GeminiClient;
@@ -20,7 +20,7 @@ export class AnswerRefiner {
       throw new Error("API key is required");
     }
     this.geminiClient = new GeminiClient(config);
-    this.sheetsService = GoogleSheetsServiceFactory.getSheetsService();
+    this.sheetsService = GoogleServiceFactory.getSheetsService();
   }
 
   async extractAnswerData(
